# swagger.yaml
openapi: 3.0.3
info:
  title: Del-Pick API
  description: API for Del-Pick, a food delivery service application enabling customers to order food, drivers to deliver orders, and store owners to manage their menus and orders.
  version: 1.0.0
  contact:
    name: API Support
    email: support@delpick.id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://delpick.horas-code.my.id/api/v1
    description: Production Server
  - url: http://localhost:6100/api/v1
    description: Development Server
...
paths:
...
  /drivers:
    get:
      tags:
        - Drivers
      summary: Retrieve all drivers
      description: Retrieves a paginated list of all drivers. Accessible to authenticated users.
      operationId: getAllDrivers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Drivers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverListResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Drivers
      summary: Create a new driver
      description: Creates a new driver account. Requires admin privileges.
      operationId: createDriver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDriverRequest'
      responses:
        '201':
          description: Driver created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                message: Driver created successfully
                data:
                  user:
                    id: 2
                    name: Jane Driver
                    email: jane@example.com
                    role: driver
                  driver:
                    id: 1
                    userId: 2
                    vehicle_number: ABC123
                    status: inactive
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /drivers/{driverId}:
    get:
      tags:
        - Drivers
      summary: Retrieve driver by ID
      description: Retrieves detailed information about a specific driver, including associated user data.
      operationId: getDriverById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/driverIdParam'
      responses:
        '200':
          description: Driver details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver details retrieved successfully
                data:
                  id: 1
                  userId: 2
                  vehicle_number: ABC123
                  status: active
                  user:
                    id: 2
                    name: Jane Driver
                    email: jane@example.com
                    role: driver
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Drivers
      summary: Update driver
      description: Updates a driver's information. Requires admin privileges.
      operationId: updateDriver
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/driverIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverRequest'
      responses:
        '200':
          description: Driver updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver updated successfully
                data:
                  id: 1
                  userId: 2
                  vehicle_number: XYZ789
                  status: active
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Drivers
      summary: Delete driver
      description: Deletes a driver and associated user account. Requires admin privileges.
      operationId: deleteDriver
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/driverIdParam'
      responses:
        '200':
          description: Driver deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver deleted successfully
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /drivers/location:
    put:
      tags:
        - Drivers
      summary: Update driver location
      description: Updates the real-time location of the authenticated driver. Requires driver privileges.
      operationId: updateDriverLocation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverLocationRequest'
      responses:
        '200':
          description: Driver location updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver location updated successfully
                data:
                  latitude: -6.12345678
                  longitude: 106.12345678
                  updatedAt: "2025-04-29T09:16:52.000Z"
        '400':
          description: Invalid location data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not a driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /drivers/{driverId}/location:
    get:
      tags:
        - Drivers
      summary: Retrieve driver location
      description: Retrieves the current location of a specific driver. Accessible to authenticated users (e.g., admins or customers tracking an order).
      operationId: getDriverLocation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/driverIdParam'
      responses:
        '200':
          description: Driver location retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver location retrieved successfully
                data:
                  driverId: 1
                  name: Jane Driver
                  latitude: -6.12345678
                  longitude: 106.12345678
                  lastUpdated: "2025-04-29T09:16:52.000Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found or location not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /drivers/status:
    put:
      tags:
        - Drivers
      summary: Update driver status
      description: Updates the status of the authenticated driver to active or inactive.
      operationId: updateDriverStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Driver status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver status updated successfully
                data:
                  id: 1
                  userId: 2
                  vehicle_number: ABC123
                  status: active
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not a driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /drivers/orders:
    get:
      tags:
        - Drivers
      summary: Retrieve driver orders
      description: Retrieves a paginated list of orders assigned to the authenticated driver, optionally filtered by status.
      operationId: getDriverOrders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/orderStatusParam'
      responses:
        '200':
          description: Driver orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverOrdersResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not a driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Driver Request Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    orderStatusParam:
      name: status
      in: query
      description: Filter by order status
      required: false
      schema:
        type: string
        enum: [pending, approved, preparing, on_delivery, delivered, cancelled]
...
    driverIdParam:
      name: driverId
      in: path
      description: ID of the driver
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
      required:
        - statusCode
        - message
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: An error occurred
        errors:
          type: string
          nullable: true
          example: Detailed error message
      required:
        - statusCode
        - message

    # Authentication Schemas
...
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [active, inactive]
          example: active

    # Customer Schemas
...
    Driver:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 2
        vehicle_number:
          type: string
          example: ABC123
        rating:
          type: number
          format: float
          example: 4.8
          nullable: true
        reviews_count:
          type: integer
          example: 50
          nullable: true
        latitude:
          type: number
          format: float
          example: -6.12345678
          nullable: true
        longitude:
          type: number
          format: float
          example: 106.12345678
          nullable: true
        status:
          type: string
          enum: [active, inactive]
          example: active
        user:
          $ref: '#/components/schemas/User'
          nullable: true
    DriverListResponse:
      type: object
      properties:
        totalItems:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 2
        currentPage:
          type: integer
          example: 1
        drivers:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
    CreateDriverRequest:
      type: object
      required:
        - name
        - email
        - password
        - phone
        - vehicle_number
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: Jane Driver
        email:
          type: string
          format: email
          maxLength: 100
          example: jane@example.com
        password:
          type: string
          minLength: 6
          maxLength: 100
          example: password123
        phone:
          type: string
          pattern: '^[0-9]{10,15}$'
          example: "08123456789"
        vehicle_number:
          type: string
          maxLength: 20
          example: ABC123
    UpdateDriverRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: Jane Driver Updated
          nullable: true
        email:
          type: string
          format: email
          maxLength: 100
          example: jane.updated@example.com
          nullable: true
        phone:
          type: string
          pattern: '^[0-9]{10,15}$'
          example: "08987654321"
          nullable: true
        vehicle_number:
          type: string
          maxLength: 20
          example: XYZ789
          nullable: true
        latitude:
          type: number
          format: float
          example: -6.12445678
          nullable: true
        longitude:
          type: number
          format: float
          example: 106.12445678
          nullable: true
        status:
          type: string
          enum: [active, inactive]
          example: active
          nullable: true
    UpdateDriverLocationRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: float
          example: -6.12345678
        longitude:
          type: number
          format: float
          example: 106.12345678
...
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        deliveryAddress:
          type: string
          example: 123 Customer Street
        subtotal:
          type: number
          format: float
          example: 10.99
        serviceCharge:
          type: number
          format: float
          example: 2.00
        total:
          type: number
          format: float
          example: 12.99
        order_status:
          type: string
          enum: [pending, approved, preparing, on_delivery, delivered, cancelled]
          example: pending
        delivery_status:
          type: string
          enum: [waiting, picking_up, on_delivery, delivered]
          example: waiting
          nullable: true
        orderDate:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"
        notes:
          type: string
          example: Please deliver quickly
          nullable: true
        customerId:
          type: integer
          example: 1
        driverId:
          type: integer
          example: 1
          nullable: true
        storeId:
          type: integer
          example: 1
        estimatedDeliveryTime:
          type: string
          format: date-time
          example: "2025-04-29T10:00:00.000Z"
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        store:
          $ref: '#/components/schemas/Store'
          nullable: true
        customer:
          $ref: '#/components/schemas/User'
          nullable: true
        driver:
          $ref: '#/components/schemas/User'
          nullable: true
...
    DriverOrdersResponse:
      type: object
      properties:
        totalItems:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 1
        currentPage:
          type: integer
          example: 1
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
...